### 2019-11-01 #1854 @whedon
[PRE REVIEW]: lightr: import spectral data and metadata in R
**Submitting author:** @bisaloo (<a href="http://orcid.org/0000-0002-4094-1476">Hugo Gruson</a>)
**Repository:** <a href="https://www.github.com/ropensci/lightr" target ="_blank">https://www.github.com/ropensci/lightr</a>
**Version:** v0.1
