### 2019-11-01 #1854 @Kevin-Mattheus-Moerman
Note: This package has been peer-reviewed by rOpenSci (v. 0.1): https://github.com/ropensci/software-review/issues/267
