### 2019-11-01 #1857 @danielskatz
Note: This package has been peer-reviewed by rOpenSci (v. 0.1): https://github.com/ropensci/software-review/issues/267
