### 2019-11-01 cfdf8e1
mostly style changes
