+++ inst/paper.md
@@ -29,10 +29,10 @@ bibliography: lightr.bib
 
 Living organisms wildly differ in their ability to see colours [@Osorio2008]. 
 For this reason, colour science relies on the use of objective measurements of
-reflectance, transmittance or aborbance spectra rather than human vision
+reflectance, transmittance, or aborbance spectra rather than human vision
 [@Bennett1994;@Cuthill1999;@Eaton2005]. These spectra are then used in vision
-models which allow scientists to predict how a given object is seen through the
-eyes of a given species (e.g. how a male bird is seen by a potential mate). This
+models that allow scientists to predict how a given object is seen through the
+eyes of a given species (e.g., how a male bird is seen by a potential mate). This
 is the basis of all studies in for example the study of the evolution of colours
 of animals and plants as communication signals
 
@@ -44,14 +44,14 @@ different formats to store spectral data. This use of non-standard file formats
 jeopardises scientific reproducibility [@Peng2009] as other researchers might
 not have the (paid) tools to open these files, and it makes us dependent on a
 third-party which might vanish anytime, leaving a trove of scientific data
-impossible to access. Vendors proprietary software sometimes have an option to
-convert those formats into human readable files such as `csv` but such softwares
+impossible to access. Vendors' proprietary software sometimes have an option to
+convert those formats into human readable files such as `csv` but such software
 are often expensive and they discard most metadata in the process. Yet, those
 metadata are critical to ensure reproducibility of the measurements, and
 ultimately of the scientific findings [@White2015].
 
-In this article, we present `lightr`, an R package which aims at offering a
-unified user-friendly interface for users to read reflectance, transmittance, 
+In this article, we present `lightr`, an R package that aims at offering a
+unified user-friendly interface for users to read reflectance, transmittance, and
 absorbance spectra files from various formats in a single line of code. 
 Additionally, it provides for the first time a fully free and open source
 solution to read proprietary spectra file formats on all operating systems.
@@ -76,10 +76,10 @@ derived from `csv` or `tsv`. Specialized parsers should usually be preferred to
 files for a single study can quickly increase. To ensure easy and efficient 
 processing of those files, `lightr` also provides three high-levels functions
 that can recursively find files and process them with a parallelized loop using
-the `parallel` R package: `lr_get_spec()` and `lr_get_metadata()` which import
+the `parallel` R package: `lr_get_spec()` and `lr_get_metadata()`, which import
 respectively spectral data and metadata as `data.frame` in R, as well as 
-`lr_convert_tocsv()` which convert all spectra files in a given folder as `csv`,
-which the same filename (minus the file extension).
+`lr_convert_tocsv()`, which converts all spectra files in a given folder as `csv`,
+with the same filename (minus the file extension).
 
 ```
 library(lightr)
@@ -88,11 +88,11 @@ lr_convert_tocsv(where = "yourfolder", ext = "ProcSpec")
 
 # Recommended workflow
 
-As mentioned earlier, proprietary spectrometry softwares can also export 
+As mentioned earlier, proprietary spectrometry software can also export 
 spectral data into a human-readable format (usually a kind of tabulation
 separated values, or `tsv`, with a complex header). `lightr` can read files 
 generated by this export step. We however
-**do not recommend you use the software built-in export function**,
+**do not recommend you use the software's built-in export function**,
 because it will apply possibly unwanted transformation to your data
 (interpolation and  subsetting) and may discard important metadata.
 
@@ -103,8 +103,8 @@ you use `lightr` to convert them into your preferred file format (such as
 
 # Usage and future directions
 
-`lightr` can serve as a basis to colour analysis R packages to deal with the
-file import step. Most of them can indeed only read a limited variety of file
[…]
