### 2019-11-01 #1857 @danielskatz
@Bisaloo - Can you please merge https://github.com/ropensci/lightr/pull/8

Also, the text in the paper that says
> It is for example planned to become a dependency to the vision model package `pavo`.

was not possible for me to fully parse so that I could suggest a change.  What is the subject of this sentence? Can you suggest a rephrasing of this sentence?
