### 2019-11-01 #1857 @Bisaloo
> `lightr` can serve as a basis for colour analysis R packages to deal with the
file import step (future versions of `pavo` for example will include `lightr` as a dependency). Most packages can indeed currently only read a limited variety of file formats.
Here is an example of a workflow where `lightr` is used to import the spectral data, which is then analysed with `pavo`:
>
> ```
> library(lightr)
> specs <- lr_get_spec(where = "yourfolder", ext = "ProcSpec")
>
> library(pavo)
> plot(specs, col = spec2rgb(specs))
> ```

Does that make more sense?
