### 2019-11-01 #1857 @danielskatz
yes.

Can you change
> It is for example planned to become a dependency to the vision model package pavo.

to
> Future versions of pavo, for example, will include lightr as a dependency.
