### 2019-11-01 1a1abcc
Fix manuscript according to @danielskatz review
