### 2019-11-01 #1857 @Bisaloo
done in https://github.com/ropensci/lightr/commit/1a1abcc6326b35b3da531f4454d7e13ef19faee9
