### 2019-11-01 #1857 @Bisaloo
Sorry for the delay. For some reason, I couldn't use GitHub integration feature on the rOpenSci repo so I mirrored it on my own account.

The DOI is 10.5281/zenodo.3525016
