### 2019-11-01 #1857 @danielskatz
Thanks @Bisaloo for submitting this to JOSS as well - it's all done now!
