### 2019-11-01 #1857 @Bisaloo
@danielskatz, sorry, I just realized (a couple of seconds too late) that the ORCID for Thomas White is wrong. Is there a way to update it?

The correct ORCID is 0000-0002-3976-1734
