### 2019-11-01 #1857 @danielskatz
@bisaloo - please change the .md file to fix this, then let us know it is done
@arfon - can you then regenerate the PDF and metadata
