### 2019-11-01 #1857 @Bisaloo
I already did it in https://github.com/ropensci/lightr/commit/e0e25a657357e64ce045afc45cc541e2ae3893ab.

Thanks!
