### 2019-11-01 #1857 @danielskatz
@arfon, can you regenerate the PDF and metadata and replace the old versions in the various places?
