### 2019-11-01 #1857 @arfon
> @arfon, can you regenerate the PDF and metadata and replace the old versions in the various places?

That sounds be updated now.
